requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
chromadb>=0.4.0
python-dotenv>=1.0.0
langchain-text-splitters>=0.0.1
requests>=2.31.0
